+++ muspinsim/cpp/celio.cpp
@@ -19,7 +19,7 @@ struct celio::EvolveContrib {
  * @param k Value of k used in the Trotter expansion
  * @param evol_contribs List of EvolveContrib structures containing information about the Hamiltonian contributions
  */
-void celio::evolve(size_t num_times, np_array_complex_t& psi, np_array_complex_t& sigma_mu, size_t half_dim, unsigned int k, const py::list& evol_contribs, np_array_double_t& results) {
+void celio::evolve(unsigned int num_times, np_array_complex_t& psi, np_array_complex_t& sigma_mu, size_t half_dim, unsigned int k, const py::list& evol_contribs, np_array_double_t& results) {
     py::buffer_info psi_info = psi.request();
     auto* psi_ptr = static_cast<std::complex<double>*>(psi_info.ptr);
